fix bug when adding folders that begin with './'
I was checking for names beginning with "." in order to
exclude such files, but was doing it on the raw string
from the user rather than the actual base name of the
file. The net result was that passing something like
"./tests" or "./keywords" would cause that folder to
be ignored. D'oh!