Fixes #41 - Files with duplicate names aren't handled properly
Prior to this change, rfhub didn't handle the case where there
were two libraries or resource files with identical names (eg:
featureA/keywords.robot, featureB/keywords.robot). This was caused
by rfhub using the last part of the name as part of the uri (eg:
/doc/keywords/Keyword1).

This change replaces the library or resource file name with a
"collection id". This roughly corresponds to the row id in the
database, though it's not guaranteed to be the row number, it's
only guaranteed to be unique *for the current running rfhub process*.

If a library or resource file name is unique,  you can still use
the name in a URI. If it is not, you'll need to use the number, which
rfhub can tell you when you get a list of all known resource files
and libraries.

This may not be the final solution for this problem, but my team
needed a quick fix.